refactor(app.tsx): move Portal to new file and some refactoring
