+++ src/types.ts
@@ -8,6 +8,7 @@ import {
 } from "./element/types";
 import { SHAPES } from "./shapes";
 import { Point as RoughPoint } from "roughjs/bin/geometry";
+import { SocketUpdateDataSource } from "./data";
 
 export type FlooredNumber = number & { _brand: "FlooredNumber" };
 export type Point = Readonly<RoughPoint>;
@@ -82,3 +83,7 @@ export declare class GestureEvent extends UIEvent {
   readonly rotation: number;
   readonly scale: number;
 }
+
+export type SocketUpdateData = SocketUpdateDataSource[keyof SocketUpdateDataSource] & {
+  _brand: "socketUpdateData";
+};
